pyproject.toml
[project]
name = "llm-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-chroma>=0.2.2",
    "langchain-community>=0.3.21",
    "langchain-core>=0.3.51",
    "langchain-deepseek>=0.1.3",
    "langchain-openai>=0.3.12",
    "langgraph>=0.3.29",
    "pypdf>=5.4.0",
    "python-dotenv>=1.1.0",
]
